- [Jira URL] Integrating the reporting script in the Github workflow for the existing e2e slack notifications.
A few words on what solution you chose (if it's the case) and describe the mechanism (if needed)

Describe the steps to be taken to test this feature. This includes configs to be made in local, the url of the app, credentials (if needed)

Add screenshots in case of simple visual changes (button color, text color, z-index etc) to give the reviewer an idea about what is expected, a before and after etc.

Section for the reviewer to describe what he/she tested and what works/doesn't work